+++ API/getStoreList.ts
@@ -1,4 +1,4 @@
-import { StoreCardType } from "../components/StoreCard";
+import { StoreCardType } from "../components/MainView/StoreCard";
 export interface StoreListDataType {
   stores: StoreCardType[];
 }
