+++ components/Auth/AuthButton.tsx
@@ -1,5 +1,5 @@
 import { StyleSheet, TouchableOpacity } from "react-native";
-import CustomText from "./CustomText";
+import CustomText from "../CustomText";
 
 interface AuthButtonProps {
   children: string;
