+++ components/MainView/StoreCard.tsx
@@ -2,9 +2,9 @@ import React from "react";
 import { Image, StyleSheet } from "react-native";
 import { TouchableOpacity } from "react-native";
 import { View } from "react-native";
-import { MainScreenNavigationProps } from "../pages/MainView";
-import CustomText from "./CustomText";
-import StatusToggle from "./StatusToggle";
+import { MainScreenNavigationProps } from "../../pages/MainView";
+import CustomText from "../CustomText";
+import StatusToggle from "../StatusToggle";
 
 export interface StoreCardType {
   id: number;
