+++ components/StoreDetailView/Carousel.tsx
