+++ pages/auth/LoginView.tsx
@@ -1,8 +1,8 @@
 import { useState } from "react";
 import { StyleSheet } from "react-native";
 import { View } from "react-native";
-import AuthButton from "../../components/AuthButton";
-import AuthLogo from "../../components/AuthLogo";
+import AuthButton from "../../components/Auth/AuthButton";
+import AuthLogo from "../../components/Auth/AuthLogo";
 import CommonInput from "../../components/CommonInput";
 
 export default function LoginView() {
