+++ pages/auth/OwnerSignUpView.tsx
@@ -2,8 +2,8 @@ import { useNavigation } from "@react-navigation/native";
 import { NativeStackScreenProps } from "@react-navigation/native-stack";
 import { StyleSheet, TouchableOpacity, View } from "react-native";
 import { RootStackParamList } from "../../App";
-import AuthBox from "../../components/AuthBox";
-import AuthLogo from "../../components/AuthLogo";
+import AuthBox from "../../components/Auth/AuthBox";
+import AuthLogo from "../../components/Auth/AuthLogo";
 import CustomText from "../../components/CustomText";
 import roleNavigateAdress from "../../function/roleNavigateAdress";
 import { SignUpScreenNavigationProps } from "./SignUpView";
